feat(app): add multi-select-icon quiz component
Mirrors `single-select-icon` with checkbox semantics. Adds a CSS
shim `fr-checkbox-rich` in `main.css` since DSFR 1.14 only ships
`fr-radio-rich`.
