+++ plateform/app/main.css
@@ -13,4 +13,135 @@
   --color-raised-grey-hover: var(--background-raised-grey-hover);
   --color-beige-gris-galet: var(--beige-gris-galet-925-125);
   --color-title-grey: var(--text-title-grey);
+  --color-blue-france-950: var(--blue-france-950-100);
+  --color-border-default-grey: var(--border-default-grey);
+  --color-background-default-grey-hover: var(--background-default-grey-hover);
+}
+
+.header-shadow {
+  --idle: transparent;
+  --hover: var(--background-raised-grey-hover);
+  --active: var(--background-raised-grey-active);
+  background-color: var(--background-raised-grey);
+  filter: drop-shadow(var(--raised-shadow));
+  flex-wrap: wrap;
+  padding-left: 0.25rem;
+  padding-right: 0.25rem;
+  width: 100%;
+  z-index: calc(var(--ground) + 750);
+}
+
+/* Variante "rich" pour les cases à cocher : DSFR 1.14 ne fournit que `fr-radio-rich`,
+   on reproduit la même mise en page (carte + pictogramme à droite) pour les checkboxes. */
+.fr-checkbox-rich {
+  position: relative;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  max-width: 100%;
+}
+
+.fr-checkbox-rich input[type="checkbox"] + label {
+  margin-left: 0;
+  padding: 0.75rem 1rem 0.75rem 2.75rem;
+  align-self: stretch;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  justify-content: center;
+  background-size:
+    100% 1px,
+    1px 100%,
+    100% 1px,
+    1px 100%;
+  background-repeat: no-repeat;
+  background-position:
+    0 0,
+    100% 0,
+    0 100%,
+    0 0;
+  background-image:
+    linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey)), linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey)),
+    linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey)), linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey));
+  background-color: var(--background-default-grey);
+  --hover: var(--background-default-grey-hover);
+  --active: var(--background-default-grey-active);
+}
+
+.fr-checkbox-rich input[type="checkbox"] + label::before {
+  left: 1rem;
+  top: 50%;
+  transform: translateY(-50%);
+}
+
+.fr-checkbox-rich input[type="checkbox"] + label .fr-hint-text {
+  margin-left: 0;
+}
+
+.fr-checkbox-rich input[type="checkbox"]:not(:disabled) ~ label:hover {
+  background-color: var(--hover);
+}
+
+.fr-checkbox-rich input[type="checkbox"]:not(:disabled) ~ label:active {
+  background-color: var(--active);
+}
+
+.fr-checkbox-rich input[type="checkbox"]:not(:disabled) ~ label:hover + .fr-checkbox-rich__pictogram {
+  background-color: var(--hover);
+}
+
+.fr-checkbox-rich input[type="checkbox"]:not(:disabled) ~ label:active + .fr-checkbox-rich__pictogram {
+  background-color: var(--active);
+}
+
+.fr-checkbox-rich input[type="checkbox"]:checked + label {
+  background-image:
+    linear-gradient(0deg, var(--border-active-blue-france), var(--border-active-blue-france)),
+    linear-gradient(0deg, var(--border-active-blue-france), var(--border-active-blue-france)),
+    linear-gradient(0deg, var(--border-active-blue-france), var(--border-active-blue-france)),
+    linear-gradient(0deg, var(--border-active-blue-france), var(--border-active-blue-france));
+}
+
+.fr-checkbox-rich input[type="checkbox"]:checked ~ .fr-checkbox-rich__pictogram {
+  background-image:
+    linear-gradient(0deg, var(--background-action-high-blue-france), var(--background-action-high-blue-france)),
+    linear-gradient(0deg, var(--background-action-high-blue-france), var(--background-action-high-blue-france)),
+    linear-gradient(0deg, var(--background-action-high-blue-france), var(--background-action-high-blue-france)),
+    linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey));
+}
+
+.fr-checkbox-rich__pictogram {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  margin-left: -1px;
+  padding: 0.25rem;
+  width: 5.5rem;
+  min-width: 5.5rem;
+  min-height: 5.5rem;
+  align-self: stretch;
+  pointer-events: none;
+  background-size:
+    100% 1px,
+    100% 1px,
+    1px 100%,
+    1px calc(100% - 0.5rem);
+  background-repeat: no-repeat;
+  background-position:
+    0 0,
+    0 100%,
+    100% 0,
+    0 0.25rem;
+  background-image:
+    linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey)), linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey)),
+    linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey)), linear-gradient(0deg, var(--border-default-grey), var(--border-default-grey));
+  background-color: var(--background-default-grey);
+}
+
+.fr-checkbox-rich__pictogram img,
+.fr-checkbox-rich__pictogram svg {
+  max-width: 3.5rem;
+  max-height: 3.5rem;
 }
